Convert modal-gallery to node:test
