--- packages/modal-gallery/jest.config.js
@@ -1,12 +1,0 @@
-const base = require('../../tools/jest/config.base.js');
-const pack = require('./package');
-
-module.exports = {
-    ...base,
-    transform: {
-        '^.+\\.js$': '../../tools/jest/babel-jest-wrapper.js'
-    },
-    displayName: pack.name,
-    name: pack.name,
-    testEnvironment: 'jsdom'
-};
