+++ packages/modal-gallery/package.json
@@ -23,9 +23,9 @@
   ],
   "scripts": {
     "build": "npx microbundle --name ModalGallery",
-    "test": "jest --coverage",
-    "dev": "webpack-dev-server --config example/webpack.config.js",
-    "prod": "webpack --config example/webpack.config.js --mode production",
+    "test": "node --test --experimental-test-coverage --test-coverage-include=\"src/**\" --import ../../tools/test-setup.mjs \"__tests__/unit/*.test.js\" && npx playwright test",
+    "dev": "webpack-dev-server --config tools/webpack.config.js",
+    "prod": "webpack --config tools/webpack.config.js --mode production",
     "prepublish": "npm run -s build"
   },
   "gitHead": "9042db4005411c4360ba3e8188e03333c10098f2"
